requirements.txt
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
dhis2.py==2.1.2
Django==2.2.1
django-compression-middleware==0.3.1
django-crispy-forms==1.11.2
django-datatables-view==1.19.1
django-discover-runner==1.0
django-role-permissions==3.1.1
django-simple-email-confirmation==0.70
django-tracking2==0.5.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
idna==2.10
logzero==1.7.0
numpy==1.19.5
pandas==1.1.5
Pillow==8.2.0
psycopg2==2.8.6
pyarrow==3.0.0
Pygments==2.8.1
python-dateutil==2.8.1
python-decouple==3.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
zstandard==0.15.2
